polycubed: move netlink library to utils folder
TODO: squash this with previous commit
