+++ src/polycubed/src/extiface.cpp
@@ -17,11 +17,10 @@
 #include "extiface.h"
 #include "bcc_mutex.h"
 #include "exceptions.h"
-#include "netlink.h"
+#include "extiface_tc.h"
 #include "patchpanel.h"
 #include "port.h"
-
-#include "extiface_tc.h"
+#include "utils/netlink.h"
 
 #include <iostream>
