+++ src/polycubed/src/extiface_xdp.cpp
@@ -19,8 +19,8 @@
 #include "bcc_mutex.h"
 #include "exceptions.h"
 #include "extiface_tc.h"
-#include "netlink.h"
 #include "patchpanel.h"
+#include "utils/netlink.h"
 
 #include <iostream>
