+++ src/services/pcn-iptables/src/modules/ChainForwarder.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "../../../../polycubed/src/extiface_info.h"
+//#include "../../../../polycubed/src/uitls/extiface_info.h"
 #include "../Iptables.h"
 #include "datapaths/Iptables_ChainForwarder_dp.h"
